ipam: move podCIDRPool to separate file
It will be used by the IPAM pools allocator in a subsequent
commit. This commit contains no functional changes.
